+++ doc/whats-new.rst
@@ -222,6 +222,10 @@ Enhancements
   (:issue:`576`).
   By `Stephan Hoyer <https://github.com/shoyer>`_.
 
+- Added :py:func:`~xarray.show_versions` function to aid in debugging
+  (:issue:`1485`).
+  By `Joe Hamman` <https://github.com/jhamman>`_.
+
 - Support using an existing, opened netCDF4 ``Dataset`` with
   :py:class:`~xarray.backends.NetCDF4DataStore`. This permits creating an
   :py:class:`~xarray.Dataset` from a netCDF4 ``Dataset`` that has been opened using
@@ -248,6 +252,11 @@ Enhancements
 - Support applying rolling window operations using bottleneck's moving window
   functions on data stored as dask arrays (:issue:`1279`).
   By `Joe Hamman <https://github.com/jhamman>`_.
+
+- Added new method :py:meth:`~Dataset.to_dask_dataframe` to
+  ``Dataset``, convert a dataset into a dask dataframe.
+  This allows lazy loading of data from a dataset containing dask arrays (:issue:`1462`).
+  By `James Munroe <https://github.com/jmunroe>`_.
 
 - Support reading and writing unlimited dimensions with h5netcdf (:issue:`1636`).
   By `Joe Hamman <https://github.com/jhamman>`_.